Вынес main в отдельный класс. Добавил класс обработчика команд
